Intégralité des fonctions implémentés

+++ backend/controllers/commentaireCtrl.js
@@ -1,4 +1,7 @@
 const db = require("../models");
+const jwtoken = require("jsonwebtoken");
+const { getUserId, getIsAdmin } = require("../Utils/jwt");
+const privateKey = process.env.PRIVATE_KEY;
 
 // create main model
 const Commentaire = db.commentaires;
@@ -10,7 +13,7 @@ const User = db.users;
 
 const addCommentaire = async (req, res) => {
   console.log("#####################################################");
-  // console.log(req.body);
+  console.log(req.body);
   console.log("#####################################################");
   let info = {
     contenu: req.body.contenu,
@@ -18,14 +21,8 @@ const addCommentaire = async (req, res) => {
     idPOST: req.body.idPOST,
   };
 
-  if (req.file) {
-    info.photo = `${req.protocol}://${req.get("host")}/images/${
-      req.file.filename
-    }`;
-  }
-
   console.log("#####################################################");
-  // console.log(info);
+  console.log(info);
   console.log("#####################################################");
 
   const commentaire = await Commentaire.create(info);
@@ -41,31 +38,33 @@ const getAllCommentaire = async (req, res) => {
   res.status(200).send(commentaires);
 };
 
-// Récuperer un post
-const getOneCommentaire = async (req, res) => {
-  const id = req.params.id;
-  const commentaire = await Commentaire.findOne({ where: { id: id } });
-  res.status(200).send(commentaire);
-};
-
-// Modifier un post
-const updateCommentaire = async (req, res) => {
-  const id = req.params.id;
-  const commentaire = await Commentaire.update(req.body, { where: { id: id } });
-  res.status(200).send(commentaire);
-};
-
 // Supprimer un Post
 const deleteCommentaire = async (req, res) => {
   const id = req.params.id;
-  await Commentaire.destroy({ where: { id: id } });
-  res.status(200).send("Le commentaire a été supprimé");
+
+  // Sécurisation de la route
+
+  const token = req.headers.cookie.split("=")[1];
+
+  const userId = getUserId(token);
+  const isAdmin = getIsAdmin(token);
+
+  const findComment = await Commentaire.findOne(req.body, {
+    where: { id: id },
+  });
+
+  if (userId === findComment.idUSER || isAdmin) {
+    await Commentaire.destroy({ where: { id: id } });
+    res.status(200).send("Le commentaire a été supprimé");
+  } else {
+    res.status(401).json({ error: "Aucun droit pour le faire" });
+  }
+
+  // TODO : Vérifier que le propriétaire est bien celui qui essaie de supprimer le commentaire (OU ADMIN)
 };
 
 module.exports = {
   addCommentaire,
   getAllCommentaire,
-  getOneCommentaire,
-  updateCommentaire,
   deleteCommentaire,
 };
